requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
df2gspread==1.0.4
et-xmlfile==1.1.0
google-api-python-client==1.6.7
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygame==2.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==3.0.1
urllib3==2.2.3
